rework tcp server + service files
